+++ lib/src/_application/workflow/workflow_service.dart
@@ -12,6 +12,11 @@ class WorkflowService {
 
   Stream<bool> get workflowStream => _workflowController.stream;
 
+  static StreamController<bool> _sectionsResetController =
+    new StreamController<bool>.broadcast();
+
+  Stream<bool> get sectionsResetStream => _sectionsResetController.stream;
+
   /// The [WorkflowService] constructor...
   WorkflowService();
 
@@ -29,4 +34,10 @@ class WorkflowService {
 
   /// The [markPreviousContentHandled] method...
   void markPreviousContentHandled() => _workflowController.add (true);
+
+  /// The [markSectionsReset] method...
+  void markSectionsReset() {
+    _workflowController.add (false);
+    _sectionsResetController.add (true);
+  }
 }